closes #620 Unused warnings fixed
